pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "circuitrl"
version = "0.1.0"
description = "Deterministic circuit RL benchmark scaffold"
requires-python = ">=3.10"
dependencies = [
  "openenv-core>=0.2.0",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "pydantic>=2.6",
  "openai>=2.0",
  "python-dotenv>=1.0",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["client", "inference", "models"]

[tool.setuptools.packages.find]
include = ["server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
